danh sách tài khoản đã xóa

+++ src/main/java/com/example/web/controller/admin/UserController/Delete.java
@@ -2,6 +2,7 @@ package com.example.web.controller.admin.UserController;
 import com.example.web.controller.util.CheckPermission;
 import com.example.web.dao.model.User;
 import com.example.web.service.UserService;
+import com.example.web.utils.SessionManager;
 import com.google.gson.Gson;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
@@ -46,8 +47,33 @@ public class Delete extends HttpServlet {
         System.out.println(id);
 
         try {
+            if (user.getId() == Integer.parseInt(id)) {
+                jsonResponse.put("status", "error");
+                jsonResponse.put("message", "Bạn không thể tự xóa chính mình!");
+                out.print(gson.toJson(jsonResponse));
+                out.flush();
+                return;
+            }
+
+            User userToDelete = userService.getUser(Integer.parseInt(id));
+            boolean isAdmin = userToDelete.getRoles().stream()
+                    .anyMatch(role -> "ADMIN".equalsIgnoreCase(role.getName()));
+            if (isAdmin) {
+                jsonResponse.put("status", "error");
+                jsonResponse.put("message", "Không thể xóa người dùng có quyền ADMIN!");
+                out.print(gson.toJson(jsonResponse));
+                out.flush();
+                return;
+            }
             boolean isDeleted = userService.deleteUser(Integer.parseInt(id));
             if (isDeleted) {
+                User up = userService.getUser(Integer.parseInt(id));
+
+                HttpSession userSession = SessionManager.userSessions.get(up.getId()+"");
+                if (userSession != null) {
+                    userSession.invalidate();
+                    SessionManager.userSessions.remove(up.getEmail());
+                }
                 jsonResponse.put("status", "success");
                 jsonResponse.put("message", "Xóa người dùng thành công!");
             } else {
